move chain related constants from util to chain package
