+++ tests/e2e/chain/config.go
@@ -1,4 +1,4 @@
-package util
+package chain
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
